Organising replies better

+++ api/migrations/20210128135306_schema.sql
@@ -1,20 +1,20 @@
 create table "source" (
-  id serial primary key,
+  id bigserial primary key,
   title text not null,
   canonical_url text not null unique,
   created_at timestamp with time zone not null default now()
 );
 
 create table "alternative" (
-  id serial primary key,
-  source_id int not null,
+  id bigserial primary key,
+  source_id bigint not null,
   url text not null unique
 );
 
 create table "opinion" (
-  id serial primary key,
-  user_id int not null,
-  source_id int not null,
+  id bigserial primary key,
+  user_id bigint not null,
+  source_id bigint not null,
   position boolean not null,
   body text not null,
   created_at timestamp with time zone not null default now()
@@ -23,17 +23,17 @@ create table "opinion" (
 create index opinion_source_id_idx on "opinion" (source_id);
 
 create table "user" (
-  id serial primary key,
+  id bigserial primary key,
   username text not null unique,
   name text not null,
   created_at timestamp with time zone not null default now()
 );
 
 create table "vote" (
-  id serial primary key,
-  user_id int not null,
-  source_id int not null,
-  opinion_id int not null,
+  id bigserial primary key,
+  user_id bigint not null,
+  source_id bigint not null,
+  opinion_id bigint not null,
   position boolean not null,
   created_at timestamp with time zone not null default now(),
   unique (source_id, user_id)
